requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0
django-dotenv==1.4.2
django-rest-auth==0.9.5
djangorestframework==3.13.1
Pillow==8.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.21
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
